Bump metallb version: Enable webhooks by default (#206)
* Makefile: Enable webhooks by default

The webhooks are now enabled by default
by MetalLB controller.

* Bump metallb version

We now use cert-controller instead of cert-manager
in MetalLB controller to create a secret and patch
the webhooks.

Now that we don't have the cert-manager dependency
for webhooks, the webhooks are enabled by default.

* CI: Remove cert-manager deployment

Deploying cert-manager for MetalLB webhooks is
not needed anymore. We now use cert-controller
in MetalLB controller to create a secret and patch
the webhooks.

* Makefile: Remove cert-manager deployment

Deploying cert-manager for MetalLB webhooks is
not needed anymore. We now use cert-controller
in MetalLB controller to create a secret and patch
the webhooks.

* Controller: Remove cert-manager permissions

cert-manager permissions for MetalLB webhooks are
not needed anymore. We now use cert-controller
in MetalLB controller to create a secret and patch
the webhooks.

* README: Remove the dependency from cert-manager

Deploying cert-manager for MetalLB webhooks is
not needed anymore. We now use cert-controller
in MetalLB controller to create a secret and patch
the webhooks.

* Bundle: Remove ENABLE_WEBHOOK var

The webhooks are now enabled by default.
No need to set the variable anymore.

* Controller: Do not create webhook secret if exist

If the webhook secret exist (was previously created,
in case of an upgrade) we should not try to override
it with an empty secret. The permissions are needed
for the controller to check the secret.

* CI: Remove ENABLE_OPERATOR_WEBHOOK var

The webhooks are now enabled by default.
No need to set the variable anymore.

* Controller: Remove ENABLE_WEBHOOK var

The webhooks are now enabled by default.
No need to set the variable anymore.

* OpenShift manifests: Disable cert-rotator

In OpenShift we don't need to use cert-rotator to
create the webhook secret. For OpenShift we use a
custom mechanism to inject the cabundle.

* CI upgrade job: Checkout working branch instead of main

We need to test upgrades against the tested branch
and not against main.

* Manifests: Enable conversion webhooks

The webhooks are now enabled by default
in MetalLB controller.

* Hack: Remove CRDs copy from MetalLB to bundle

We don't need to copy the CRDs from MetalLB to
the bundle manifests since this is done when
generating the bundle.

* CI: Apply patch for backward_compatible lane

The patch changes the e2e cleanup, so that the BFDProfile
will be deleted first, and BGPPeer after. This is needed
for the BFDprofile validation webhook that denies the request
to delete the BFDProfile since it is used by a BGPPeer.