Deploy the webhooks and the backend upfront (#212)
* Config: deploy the webhooks and the backend upfront

The crds are deployed as part of the operator deployment (either via the
csv or via the manifests), while the webhooks are deployed only after
the creation of the metallb resource. This allows the creation of
unchecked resources before metallb is deployed.
In order to avoid that, we deploy also the backend of the webhooks
which is a trimmed down controller in "webhook only" mode.
Also, the controller deployed by the operator will have the webhook
disabled.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Manifest generation: do not import webhook related objects

The webhooks and all the related objects are now part of the deployment,
the operator must not deploy them so we filter them.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Controller: remove the secret logic

The webhook secret is now deployed as part of the deployment, so the
controller doesn't need to deploy it (nor to apply its logic).

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Change the namespace injected env variable

Since the operator now targets all the namespaces, we won't know what
namespace the operator is deployed in. Had to change the env variable
otherwise the automation will replace the value with
olm.watchedNamespaces again.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Config: add prefix only to daemonset and deployment

When deploying via manifests, the controller - webhook expects a well
known secret name. If we add a prefix, the webhook endpoint won't work.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Align the bin general manifest

Changing the manifest to align with webhooks deployed outside of the
operator.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>