Make IsMetalLBAvailable consider ready status
IsMetalLBAvailable is the function that checks if metallb is deployed
and available, which results in the update of the metallb resource condition.

Currently, we only check the CurrentNumberScheduled, but this is doesn't take
into account a situation when the speaker pods are scheduled but are not
in the "Ready" state.
Therefore, here we add a check against ds.Status.NumberReady.

Signed-off-by: liornoy <lnoy@redhat.com>