Refactor pkg/helm contents
We rename/move some items so that there is a file which contains the stuff
that is specific to metallb's helm chart and one containing common actions.
We also refactor the With* functions to return the object and assign it to the
map from outside as it reads better.
This is in prepration for adding the frr-k8s chart that the operator will manage,
where it will be a separate file.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>