Bundle: remove the conversion section from the crd
The conversion section of the crd refers to the metallb conversion
implementation, which is done by the deployment of metallb through helm
/ manifests.

In the operator though, we let olm patch the crds based on what's in the
csv. For this reason, we remove the conversion section to avoid races
when the crd is created but olm did not patch it, preferring an empty
conversion strategy to one referring a non existing service, which would
make the webhook invocation fail.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>